[ELY-2359] Add NonceManager interface
