+++ http/digest/src/main/java/org/wildfly/security/http/digest/DigestMechanismFactory.java
@@ -28,7 +28,7 @@ import static org.wildfly.security.http.HttpConstants.DIGEST_SHA512_256_NAME;
 import static org.wildfly.security.http.HttpConstants.MD5;
 import static org.wildfly.security.http.HttpConstants.SHA256;
 import static org.wildfly.security.http.HttpConstants.SHA512_256;
-import static org.wildfly.security.http.digest.NonceManager.*;
+import static org.wildfly.security.http.digest.DefaultNonceManager.*;
 import static org.wildfly.security.provider.util.ProviderUtil.INSTALLED_PROVIDERS;
 
 import java.security.Provider;
@@ -75,7 +75,7 @@ public class DigestMechanismFactory implements HttpServerAuthenticationMechanism
      * SHA256 Digest Algorithm
      */
 
-    private static NonceManager nonceManager = new NonceManager(DEFAULT_VALIDITY_PERIOD, DEFAULT_NONCE_SESSION_TIME, true, DEFAULT_KEY_SIZE, SHA256, ElytronMessages.httpDigest);
+    private static NonceManager nonceManager = new DefaultNonceManager(DEFAULT_VALIDITY_PERIOD, DEFAULT_NONCE_SESSION_TIME, true, DEFAULT_KEY_SIZE, SHA256, ElytronMessages.httpDigest);
 
     /**
      * @see org.wildfly.security.http.HttpServerAuthenticationMechanismFactory#getMechanismNames(java.util.Map)
@@ -101,10 +101,10 @@ public class DigestMechanismFactory implements HttpServerAuthenticationMechanism
         checkNotNullParam("callbackHandler", callbackHandler);
 
         if (properties.containsKey("nonceManager")) {
-            nonceManager = (NonceManager) properties.get("nonceManager");
+            nonceManager = (DefaultNonceManager) properties.get("nonceManager");
         } else if (properties.get(CONFIG_SESSION_BASED_DIGEST_NONCE_MANAGER) != null) {
             if (Boolean.parseBoolean((String) properties.get(CONFIG_SESSION_BASED_DIGEST_NONCE_MANAGER))) {
-                nonceManager = new NonceManager(DEFAULT_VALIDITY_PERIOD, DEFAULT_NONCE_SESSION_TIME, true, DEFAULT_KEY_SIZE, SHA256, ElytronMessages.httpDigest, null, true);
+                nonceManager = new DefaultNonceManager(DEFAULT_VALIDITY_PERIOD, DEFAULT_NONCE_SESSION_TIME, true, DEFAULT_KEY_SIZE, SHA256, ElytronMessages.httpDigest, null, true);
             }
         }
